Added authorization for all controllers

+++ ProjectLedg.Server/Controllers/IngoingInvoiceController.cs
@@ -4,7 +4,10 @@ using Microsoft.AspNetCore.Mvc;
 using ProjectLedg.Server.Data.Models;
 using ProjectLedg.Server.Data.Models.DTOs.Invoice;
 using ProjectLedg.Server.Helpers.File_Compressor;
+using ProjectLedg.Server.Services;
 using ProjectLedg.Server.Services.IServices;
+using System.ComponentModel.Design;
+using System.Security.Claims;
 
 namespace ProjectLedg.Server.Controllers
 {
@@ -15,16 +18,19 @@ namespace ProjectLedg.Server.Controllers
         private readonly IIngoingInvoiceService _invoiceService;
         private readonly IFormRecognizerService _formService;
         private readonly IBlobStorageService _blobStorageService;
+        private readonly IUserService _userService;
         private readonly ILogger<IngoingInvoiceController> _logger;
 
-        public IngoingInvoiceController(IIngoingInvoiceService invoiceService, IFormRecognizerService formService, IBlobStorageService blobStorageService, ILogger<IngoingInvoiceController> logger)
+        public IngoingInvoiceController(IIngoingInvoiceService invoiceService, IFormRecognizerService formService, IBlobStorageService blobStorageService, ILogger<IngoingInvoiceController> logger, IUserService userService)
         {
             _invoiceService = invoiceService;
             _formService = formService;
             _blobStorageService = blobStorageService;
             _logger = logger;
+            _userService = userService;
         }
 
+        [Authorize]
         [HttpPost("analyze")]
         public async Task<IActionResult> AnalyzeForm(IFormFile file)
         {
@@ -89,14 +95,22 @@ namespace ProjectLedg.Server.Controllers
             }
         }
 
-
-
-
         [Authorize]
         [HttpPost("save")]
         public async Task<IActionResult> SaveIngoingInvoice([FromBody] InvoiceDTO invoiceDto)
-
         {
+            // Get claims from JWT and user id from that
+            ClaimsPrincipal userClaims = User;
+            string userId = userClaims.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return Unauthorized("No user id in claims.");
+
+            // Verify that the company we want data for/from belongs to this user
+            var companyBelongsToUser = await _userService.VerifyCompanyBelongsToUser(userId, invoiceDto.CompanyId);
+            if (!companyBelongsToUser.Success)
+                return BadRequest(companyBelongsToUser.Message);
+
+
             var tempFilePath = HttpContext.Session.GetString("TempFilePath");
 
             if (string.IsNullOrEmpty(tempFilePath) || !System.IO.File.Exists(tempFilePath))
@@ -104,32 +118,27 @@ namespace ProjectLedg.Server.Controllers
                 return BadRequest("Temporary file not found. Please upload the file again.");
             }
 
-            //var userId = User.Claims.FirstOrDefault(c => c.Type == "sub")?.Value;
-            //if (string.IsNullOrEmpty(userId))
-            //{
-            //    return Unauthorized("User not found in token.");
-            //}
-
-            var userId = "testUser"; 
             var result = await _invoiceService.SaveIngoingInvoiceAsync(invoiceDto, tempFilePath, userId);
 
             return Ok(new { message = "Invoice saved successfully!" });
         }
 
-        [HttpGet("{invoiceId}")]
-        public async Task<IActionResult> GetInvoice(int invoiceId)
-        {
-            var invoice = await _invoiceService.GetIngoingInvoiceByIdAsync(invoiceId);
-            if (invoice == null)
-            {
-                return NotFound("Invoice not found.");
-            }
-            return Ok(invoice);
-        }
-
+        [Authorize]
         [HttpGet("all/Company/{companyId}")]
         public async Task<IActionResult> GetAllIngoingInvoicesForCompanyAsync(int companyId)
         {
+            // Get claims from JWT and user id from that
+            ClaimsPrincipal userClaims = User;
+            string userId = userClaims.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return Unauthorized("No user id in claims.");
+
+            // Verify that the company we want data for/from belongs to this user
+            var companyBelongsToUser = await _userService.VerifyCompanyBelongsToUser(userId, companyId);
+            if (!companyBelongsToUser.Success)
+                return BadRequest(companyBelongsToUser.Message);
+
+
             try
             {
                 var invoices = await _invoiceService.GetAllIngoingInvoicesForCompanyAsync(companyId);
@@ -146,6 +155,8 @@ namespace ProjectLedg.Server.Controllers
             
         }
 
+        // Only accessible for Admins and not users - there already exists methods and endpoints to get all and specific invoices for companies
+        [Authorize(Roles = "Manager, Admin")]
         [HttpGet("all")]
         public async Task<IActionResult> GetAllInvoices()
         {
@@ -156,9 +167,45 @@ namespace ProjectLedg.Server.Controllers
             return Ok(invoices);
         }
 
+
+        [Authorize]
+        [HttpGet("{invoiceId}")]
+        public async Task<IActionResult> GetInvoice(int invoiceId)
+        {
+            // Get claims from JWT and user id from that
+            ClaimsPrincipal userClaims = User;
+            string userId = userClaims.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return Unauthorized("No user id in claims.");
+
+            // Verify that the invoice we want to access belongs to this user
+            var invoiceBelongsToUser = await _userService.VerifyInvoiceBelongsToUserAsync(userId, invoiceId);
+            if (!invoiceBelongsToUser.Success)
+                return BadRequest(invoiceBelongsToUser.Message);
+
+            var invoice = await _invoiceService.GetIngoingInvoiceByIdAsync(invoiceId);
+            if (invoice == null)
+            {
+                return NotFound("Invoice not found.");
+            }
+            return Ok(invoice);
+        }
+
+        [Authorize]
         [HttpPost("create")]
         public async Task<IActionResult> CreateInvoice([FromBody] InvoiceDTO invoiceDto)
         {
+            // Get claims from JWT and user id from that
+            ClaimsPrincipal userClaims = User;
+            string userId = userClaims.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return Unauthorized("No user id in claims.");
+
+            // Verify that the company we want data for/from belongs to this user
+            var companyBelongsToUser = await _userService.VerifyCompanyBelongsToUser(userId, invoiceDto.CompanyId);
+            if (!companyBelongsToUser.Success)
+                return BadRequest(companyBelongsToUser.Message);
+
             var result = await _invoiceService.CreateIngoingInvoiceAsync(invoiceDto);
             if (result == null)
             {
@@ -168,9 +215,26 @@ namespace ProjectLedg.Server.Controllers
             return Ok(new { message = "Invoice created successfully!" });
         }
 
+        [Authorize]
         [HttpPut("update/{invoiceId}")]
         public async Task<IActionResult> UpdateInvoice(int invoiceId, [FromBody] InvoiceDTO invoiceDto)
         {
+            // Get claims from JWT and user id from that
+            ClaimsPrincipal userClaims = User;
+            string userId = userClaims.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return Unauthorized("No user id in claims.");
+
+            // Verify that the company we want data for/from belongs to this user
+            var companyBelongsToUser = await _userService.VerifyCompanyBelongsToUser(userId, invoiceDto.CompanyId);
+            if (!companyBelongsToUser.Success)
+                return BadRequest(companyBelongsToUser.Message);
+
+            // Verify that the invoice we want to delete belongs to this user
+            var invoiceBelongsToUser = await _userService.VerifyInvoiceBelongsToUserAsync(userId, invoiceId);
+            if (!invoiceBelongsToUser.Success)
+                return BadRequest(invoiceBelongsToUser.Message);
+
             var result = await _invoiceService.UpdateIngoingInvoiceAsync(invoiceId, invoiceDto);
             if (!result)
             {
@@ -180,9 +244,21 @@ namespace ProjectLedg.Server.Controllers
             return Ok(new { message = "Invoice updated successfully!" });
         }
 
+        [Authorize]
         [HttpDelete("delete/{invoiceId}")]
         public async Task<IActionResult> DeleteInvoice(int invoiceId)
         {
+            // Get claims from JWT and user id from that
+            ClaimsPrincipal userClaims = User;
+            string userId = userClaims.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userId == null)
+                return Unauthorized("No user id in claims.");
+
+            // Verify that the invoice we want to delete belongs to this user
+            var invoiceBelongsToUser = await _userService.VerifyInvoiceBelongsToUserAsync(userId, invoiceId);
+            if (!invoiceBelongsToUser.Success)
+                return BadRequest(invoiceBelongsToUser.Message);
+
             var result = await _invoiceService.DeleteIngoingInvoiceAsync(invoiceId);
             if (!result)
             {
